Implement tree-based resolution of StaticResources
This replaces the previous approach (only a single copy of a resource was possible/permitted per app).

Uno's new approach doesn't exactly match UWP (parse-time XAML scope resolution), instead it uses a 'two-speed' approach where the lookup is made in App.Resources when the XAML is initialized, and then again at load-time based on the resources in scope in the local tree.